# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almáskosár API",
        "description": "Az Almáskosár Webshop API végpontjai",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Listázza az összes megrendelést",
                "description": "Listázza az összes megrendelést",
                "operationId": "getOrdersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrdersSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orders not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Új rendelés létrehozása",
                "description": "Új rendelés létrehozása",
                "operationId": "33d737c93193573b1687e463b1562642",
                "requestBody": {
                    "description": "Új rendelés adatai",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrdersSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A rendelés sikeresen létrejött",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Az üzenet",
                                            "type": "string",
                                            "example": "Sikeresen hozzáadva!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/edit": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Rendelés keresése",
                "description": "Visszaadja a keresett rendelés adatait",
                "operationId": "5ffbeb300bf10a3eb9310ff991f21ac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Keresett rendelés azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visszaadja a keresett rendelést",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrdersSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Rendelés szerkesztése",
                "description": "Frissíti a keresett rendelés adatait",
                "operationId": "updateOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Frissíteni kívánt rendelés azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrdersSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrdersSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Rendelés törlése",
                "description": "Törli a keresett rendelést",
                "operationId": "3bdb01ae061166f8fcc7a9f60544a80a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Törölni kívánt rendelés azonosítója",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrdersSchema": {
                "properties": {
                    "id": {
                        "description": "Rendelés azonosítója",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "user_id": {
                        "description": "Felhasználó azonosítója",
                        "type": "integer",
                        "example": "5",
                        "nullable": false
                    },
                    "last_name": {
                        "description": "Felhasználó vezetékneve",
                        "type": "string",
                        "example": "Gipszes",
                        "nullable": true
                    },
                    "first_name": {
                        "description": "Felhasználó keresztneve",
                        "type": "string",
                        "example": "Jakab",
                        "nullable": true
                    },
                    "email": {
                        "description": "Felhasználó email címe",
                        "type": "string",
                        "format": "email",
                        "example": "gipszes.jakab@gmail.com",
                        "nullable": false
                    },
                    "postal_code": {
                        "description": "Felhasználó irányítószáma",
                        "type": "integer",
                        "example": "1454",
                        "nullable": false
                    },
                    "city": {
                        "description": "Felhasználó városa",
                        "type": "string",
                        "example": "Budapest",
                        "nullable": false
                    },
                    "address": {
                        "description": "Felhasználó címe",
                        "type": "string",
                        "example": "Lilaakác u. 15.",
                        "nullable": false
                    },
                    "phone_number": {
                        "description": "Felhasználó telefonszáma",
                        "type": "string",
                        "example": "+36302124787",
                        "nullable": false
                    },
                    "customer_code": {
                        "description": "Felhasználó vevőkódja",
                        "type": "string",
                        "example": "84222",
                        "nullable": false
                    },
                    "total_price": {
                        "description": "Vásárlás végösszege",
                        "type": "string",
                        "example": "929999",
                        "nullable": false
                    },
                    "status": {
                        "description": "Vásárlás státusza",
                        "type": "string",
                        "example": "0",
                        "nullable": false
                    },
                    "message": {
                        "description": "Rendeléssel kapcsolatos információk a felhasználónak",
                        "type": "string",
                        "example": "Rendelés kiszállítás alatt",
                        "nullable": true
                    },
                    "tracking_number": {
                        "description": "Rendelés száma",
                        "type": "string",
                        "example": "Almáskosár/1339",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    }
}